+++ include/strings.h
@@ -44,9 +44,12 @@ extern const u8 gText_NuzDeleteOnwhiteoutOn[];
 extern const u8 gText_NuzDeleteOnwhiteoutOff[];
 extern const u8 gText_NuzSingleCatchOn[];
 extern const u8 gText_NuzSingleCatchOff[];
+extern const u8 gText_NuzStarterChooseOn[];
+extern const u8 gText_NuzStarterChooseOff[];
 extern const u8 gText_FaintDelete[];
 extern const u8 gText_WODelete[];
 extern const u8 gText_SingleCatch[];
+extern const u8 gText_StarterChoose[];
 
 extern const u8 gText_Normal[];
 extern const u8 gText_Super[];